1. add private to LoginPage and BurgerConstructorPage 2. public void setup() add to TestBase.java
