+++ src/main/java/pages/LoginPage.java
@@ -6,7 +6,7 @@ import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 
 public class LoginPage {
-    public final By textPutBurger = By.xpath("//h1[text()='Соберите бургер']");
+    private final By textPutBurger = By.xpath("//h1[text()='Соберите бургер']");
     private final WebDriver driver;
     private final By loginAccButton = By.xpath("//button[text()='Войти в аккаунт']");
     private final By emailField = By.xpath("//input[@name='name']");
